Melhorias NMAR
