+++ Strategy/MissingDataStrategy.py
@@ -1,13 +1,15 @@
 from abc import ABC, abstractmethod
+from multiprocessing.sharedctypes import Value
 from numpy import NaN
 import pandas as pd
+import re
 
 class IMissingDataStrategy(ABC):
     """
     Interface for implementing missing data mechanisms with strategy pattern
     """
     @abstractmethod
-    def execute(self, attribute, missing_rate) -> None:
+    def execute(self, attribute: pd.Series, missing_rate: float, query: str) -> None:
         pass
 
 
@@ -15,7 +17,7 @@ class MCAR(IMissingDataStrategy):
     """
     MCAR mechanism strategy
     """
-    def execute(self, column: pd.Series, missing_rate: float) -> pd.Series:
+    def execute(self, column: pd.Series, missing_rate: float, query: str) -> pd.Series:
         """
         Erase values from a series with MCAR pattern
         """
@@ -27,7 +29,7 @@ class MAR(IMissingDataStrategy):
     """
     MAR mechanism strategy
     """
-    def execute(self, column: pd.Series, missing_rate: float) -> pd.Series:
+    def execute(self, column: pd.Series, missing_rate: float, query: str) -> pd.Series:
         """
         Erase values from a series with MAR pattern
         """
@@ -38,41 +40,24 @@ class NMAR(IMissingDataStrategy):
     """
     NMAR mechanism strategy
     """
-
-    def input_operator(self) -> str:
+    def execute(self, column: pd.Series, missing_rate: float, query: str) -> pd.Series:
         """
-        Input and validate an operator
+        Erase values from 'column' with NMAR pattern, filtering column by 'query' at a rate of 'missing_rate'
         """
-        operator = input("Type in one of the following operators: ==, >, >=, <, <=\n")
+        df = pd.DataFrame({'x': column})
+        trimmed_query = re.sub('[\s+]', '', query)
 
-        if operator not in ["==", ">", ">=", "<", "<="]:
-            raise ValueError(f"Invalid operator '{operator}'. Should be one of (==, >, >=, <, <=)")
-
-        return operator
-
-    def input_value(self) -> str:
-        """
-        Input and validate a value
-        """
-        value = input("Type in a value: ")
         try:
-            return float(value)
+            filtered_df = df.query(f"{trimmed_query}")
+
         except:
-            raise TypeError(f"Invalid value '{value}'. Should be a number.")
+            raise ValueError(f"Invalid query '{query}' for NMAR mechanism.")
 
+        else:
+            print(f"Erasing values with NMAR pattern.")
+            print(f"Query: x in {column.name} where {query}")
+            print(f"Missing rate: {missing_rate}")
 
-    def execute(self, column: pd.Series, missing_rate: float) -> pd.Series:
-        """
-        Erase values from a series with NMAR pattern
-        """
-        print("Choose a condition to apply NMAR pattern over the column\n")
-        operator = self.input_operator()
-        value = self.input_value()
-        
-        print(f"Erasing values with NMAR pattern.\n Condition: col_value {operator} {value} \n Missing rate: {missing_rate}")
+            df.loc[filtered_df.sample(frac=missing_rate).index, 'x'] = NaN
 
-        df = pd.DataFrame({'col': column})
-
-        df.loc[df.query(f"col {operator} {value}").sample(frac=missing_rate).index, 'col'] = NaN
-
-        return df.col
+            return df.x
